support Nullable<Enum> command arguments
