+++ DSharpPlus.Commands/Converters/EnumArgumentConverter.cs
@@ -14,15 +14,26 @@ public class EnumConverter : ISlashArgumentConverter<Enum>, ITextArgumentConvert
     public string ReadableName => "Multiple Choice";
     public bool RequiresText => true;
 
-    public Task<Optional<Enum>> ConvertAsync(TextConverterContext context, MessageCreateEventArgs eventArgs) => Task.FromResult(Enum.TryParse(context.Parameter.Type, context.Argument, true, out object? result)
-        ? Optional.FromValue((Enum)result)
-        : Optional.FromNoValue<Enum>());
+    public Task<Optional<Enum>> ConvertAsync(TextConverterContext context, MessageCreateEventArgs eventArgs)
+    {
+        return Task.FromResult(Enum.TryParse(context.Parameter.Type, context.Argument, true, out object? result)
+            ? Optional.FromValue((Enum)result)
+            : Optional.FromNoValue<Enum>());
+    }
 
     public Task<Optional<Enum>> ConvertAsync(InteractionConverterContext context, InteractionCreateEventArgs eventArgs)
     {
-        object value = Convert.ChangeType(context.Argument.Value, Enum.GetUnderlyingType(context.Parameter.Type), CultureInfo.InvariantCulture);
-        return Enum.IsDefined(context.Parameter.Type, value)
-            ? Task.FromResult(Optional.FromValue((Enum)Enum.ToObject(context.Parameter.Type, value)))
+        Type effectiveType = Nullable.GetUnderlyingType(context.Parameter.Type) ?? context.Parameter.Type;
+
+        object value = Convert.ChangeType
+        (
+            context.Argument.Value,
+            Enum.GetUnderlyingType(effectiveType),
+            CultureInfo.InvariantCulture
+        );
+
+        return Enum.IsDefined(effectiveType, value)
+            ? Task.FromResult(Optional.FromValue((Enum)Enum.ToObject(effectiveType, value)))
             : Task.FromResult(Optional.FromNoValue<Enum>());
     }
 }
